moves Tessel functionality into individual files
