+++ lib/controller.js
@@ -1,6 +1,6 @@
 var usb = require('./usb_connection')
   , lan = require('./lan_connection')
-  , Tessel = require('./tessel')
+  , Tessel = require('./tessel/tessel')
   , async = require('async');
   ;
 
@@ -98,7 +98,7 @@ function deployScript(opts, push, callback) {
     }
 
     // Run the script on Tessel
-    tessel.runScript(opts, push);
+    tessel.deployScript(opts, push);
   });
 }
