+++ lib/lan_connection.js
@@ -2,7 +2,7 @@ var Client = require('ssh2').Client
   , envfile = require('envfile')
   , fs = require('fs')
   , async = require('async')
-  , Tessel = require('./tessel')
+  , Tessel = require('./tessel/tessel')
   ;
 
 function LANConnection(opts) {
