+++ lib/tessel/commands.js
