+++ lib/tessel/deploy.js
@@ -1,40 +1,17 @@
-var commands = require('./commands')
+var Tessel = require('./tessel')
+  , commands = require('./commands')
   , tessel = require('tessel')
-  ;
+  ; 
 
 var RUN_PATH = '/tmp/remote-script/';
 var PUSH_PATH = '/app';
-
-/* 
-  Constructor function for Tessel objects
-    param connection: the Connection object that represents the physical comm bus
-*/
-function Tessel(connection) {
-  // Set the connection var so we have an abstract interface to relay comms
-  this.connection = connection;
-  // The human readable name of the device
-  this.name;
-  // The unique serial number of the device
-  this.serialNumber;
-}
-
-/* 
-  List available WiFi networks
-    param opts: unused so far
-    param callback: called upon completion in the form (err, null)
-*/
-Tessel.prototype.scanWiFi = function(opts, callback) {
-  // this.connection.write(commands.scanWiFi, function(err, readStream) {
-    
-  // });
-};
 
 /* 
   Run a script on a Tessel
     param opts: unused so far
     param push: whether or not this script is pushed into RAM or flash
 */
-Tessel.prototype.runScript = function(opts, push) {
+Tessel.prototype.deployScript = function(opts, push) {
   var self = this;
 
   var filepath = push ? PUSH_PATH : RUN_PATH;
@@ -109,34 +86,3 @@ Tessel.prototype.runScript = function(opts, push) {
     });
   });
 }
-
-Tessel.prototype.eraseScript = function(opts, callback) {
-  tessel.logs.info('Erasing code...');
-  // Stop processes and delete everything in the folder
-  this.connection.exec(commands.stopRunningScript(PUSH_PATH), function(err, streams) {
-    streams.stdin.on('error', function(e) {
-      tessel.logs.error("Unable to erase code:", e);
-      return callback && callback(err);
-    });
-    streams.stdout.end(function () {
-      tessel.logs.info('Code erased.');
-      return callback && Callback();
-    });
-  });
-};
-
-/* 
-  Fetch the human friendly name of this Tessel.
-    param callback: called upon completion in the form (err)
-*/
-Tessel.prototype.getName = function(callback) {
-  // TODO: actually fetch a name
-  // this.write(commands.getIdentification, function(err, readStream) {
-
-  // });
-  // This is for testing purposes only
-  this.name = 'Jon\'s Tessel'
-  callback && callback();
-};
-
-module.exports = Tessel;
